// src/Utils/interfaces/getPokemon.interface.ts
export const PokemonDetailedEmpty: PokemonDetailed = {
    abilities: [],
    base_experience: 0,
    forms: [],
    game_indices: [],
    height: 0,
    held_items: [],
    id: 0,
    is_default: false,
    location_area_encounters: '',
    moves: [],
    name: '',
    order: 0,
    species: {
        name: '',
        url: '',
    },
    sprites: {
        back_default: '',
        back_female: '',
        back_shiny: '',
        back_shiny_female: '',
        front_default: '',
        front_female: '',
        front_shiny: '',
        front_shiny_female: '',
    },
    past_types: [],
    stats: [],
    types: [],
    weight: 0,
}

export interface PokemonDetailed {
    abilities: Ability[]
    base_experience: number
    forms: Species[]
    game_indices: GameIndex[]
    height: number
    held_items: unknown[]
    id: number
    is_default: boolean
    location_area_encounters: string
    moves: Move[]
    name: string
    order: number
    past_types: unknown[]
    species: Species
    sprites: Sprites
    stats: Stat[]
    types: Type[]
    weight: number
}

export interface Ability {
    ability: Species
    is_hidden: boolean
    slot: number
}

export interface Species {
    name: string
    url: string
}

export interface GameIndex {
    game_index: number
    version: Species
}

export interface Move {
    move: Species
    version_group_details: VersionGroupDetail[]
}

export interface VersionGroupDetail {
    level_learned_at: number
    move_learn_method: Species
    version_group: Species
}

export interface GenerationV {
    'black-white': Sprites
}

export interface GenerationIv {
    'diamond-pearl': Sprites
    'heartgold-soulsilver': Sprites
    platinum: Sprites
}

export interface Versions {
    'generation-i': GenerationI
    'generation-ii': GenerationIi
    'generation-iii': GenerationIii
    'generation-iv': GenerationIv
    'generation-v': GenerationV
    'generation-vi': { [key: string]: Home }
    'generation-vii': GenerationVii
    'generation-viii': GenerationViii
}

export interface Sprites {
    back_default: string
    back_female: string
    back_shiny: string
    back_shiny_female: string
    front_default: string
    front_female: string
    front_shiny: string
    front_shiny_female: string
    other?: Other
    versions?: Versions
    animated?: Sprites
}
// ...
export interface Other {
    dream_world: DreamWorld
    home: Home
    'official-artwork': OfficialArtwork
}

export interface OfficialArtwork {
    front_default: string
}

export interface Stat {
    base_stat: number
    effort: number
    stat: Species
}

export interface Type {
    slot: number
    type: Species
}